Move `installerProvider` to `LLMService`
