+++ src/benchmark/framework/structures/llmServiceProvider/llmServiceProvider.ts
@@ -1,4 +1,5 @@
 import { ErrorsHandlingMode } from "../../../../llm/llmServices/commonStructures/errorsHandlingMode";
+import { InstallerProvider } from "../../../../llm/llmServices/commonStructures/installerProvider";
 import { LLMService } from "../../../../llm/llmServices/llmService";
 import { ModelParams } from "../../../../llm/llmServices/modelParams";
 import { ParamsResolverImpl } from "../../../../llm/llmServices/utils/paramsResolvers/paramsResolverImpl";
@@ -10,8 +11,6 @@ import {
     invariantFailed,
     throwError,
 } from "../../../../utils/errors/throwErrors";
-
-import { InstallerProvider } from "./installerProvider";
 
 export abstract class LLMServiceProvider {
     protected abstract readonly selfClass: LLMServiceProviderClass;
